@@ -29,7 +29,7 @@ with open("HISTORY.rst") as history_file:
     history = history_file.read()
 
 requirements = [
-    "dask[array] >=0.16.1",
+    "dask[array] >=2021.10.0",
     "numpy >=1.11.3",
     "scipy >=0.19.1",
     "pims >=0.4.1",
@@ -54,13 +54,12 @@ setup(
     version=versioneer.get_version(),
     description="Distributed image processing",
     long_description=readme + "\n\n" + history,
-    author="John Kirkham",
-    author_email="kirkhamj@janelia.hhmi.org",
+    author="dask-image contributors",
     url="https://github.com/dask/dask-image",
     cmdclass=cmdclasses,
     packages=setuptools.find_packages(exclude=["tests*"]),
     include_package_data=True,
-    python_requires='>=3.5',
+    python_requires='>=3.8',
     install_requires=requirements,
     license="BSD 3-Clause",
     zip_safe=False,
@@ -71,10 +70,10 @@ setup(
         "License :: OSI Approved :: BSD License",
         "Natural Language :: English",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.5",
-        "Programming Language :: Python :: 3.6",
-        "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
     ],
     tests_require=test_requirements
 )
