@@ -1,14 +1,12 @@
 # -*- coding: utf-8 -*-
 
-__author__ = """John Kirkham"""
-__email__ = "kirkhamj@janelia.hhmi.org"
-
 __all__ = [
     "convolve",
     "correlate",
     "laplace",
     "prewitt",
     "sobel",
+    "gaussian",
     "gaussian_filter",
     "gaussian_gradient_magnitude",
     "gaussian_laplace",
@@ -22,46 +20,16 @@ __all__ = [
     "threshold_local",
 ]
 
-from ._conv import (
-    convolve,
-    correlate,
-)
-
-from ._diff import (
-    laplace,
-)
-
-from ._edge import (
-    prewitt,
-    sobel,
-)
-
-from ._gaussian import (
-    gaussian_filter,
-    gaussian_gradient_magnitude,
-    gaussian_laplace,
-)
-
-from ._generic import (
-    generic_filter,
-)
-
-from ._order import (
-    minimum_filter,
-    median_filter,
-    maximum_filter,
-    rank_filter,
-    percentile_filter,
-)
-
-from ._smooth import (
-    uniform_filter,
-)
-
-from ._threshold import (
-    threshold_local,
-)
-
+from ._conv import convolve, correlate
+from ._diff import laplace
+from ._edge import prewitt, sobel
+from ._gaussian import (gaussian, gaussian_filter,
+                        gaussian_gradient_magnitude, gaussian_laplace)
+from ._generic import generic_filter
+from ._order import (maximum_filter, median_filter, minimum_filter,
+                     percentile_filter, rank_filter)
+from ._smooth import uniform_filter
+from ._threshold import threshold_local
 
 convolve.__module__ = __name__
 correlate.__module__ = __name__
@@ -71,6 +39,7 @@ laplace.__module__ = __name__
 prewitt.__module__ = __name__
 sobel.__module__ = __name__
 
+gaussian.__module__ = __name__
 gaussian_filter.__module__ = __name__
 gaussian_gradient_magnitude.__module__ = __name__
 gaussian_laplace.__module__ = __name__
