@@ -1,11 +1,10 @@
 # -*- coding: utf-8 -*-
-import scipy.ndimage.filters
+import scipy.ndimage
 
-from . import _utils
+from ..dispatch._dispatch_ndfilters import (dispatch_convolve,
+                                            dispatch_correlate)
 from ..dispatch._utils import check_arraytypes_compatible
-from ..dispatch._dispatch_ndfilters import (
-    dispatch_convolve,
-    dispatch_correlate)
+from . import _utils
 
 __all__ = [
     "convolve",
@@ -13,17 +12,17 @@ __all__ = [
 ]
 
 
-@_utils._update_wrapper(scipy.ndimage.filters.convolve)
-def convolve(image,
-             weights,
-             mode='reflect',
-             cval=0.0,
-             origin=0):
+@_utils._update_wrapper(scipy.ndimage.convolve)
+def convolve(image, weights, mode="reflect", cval=0.0, origin=0):
     check_arraytypes_compatible(image, weights)
 
     origin = _utils._get_origin(weights.shape, origin)
     depth = _utils._get_depth(weights.shape, origin)
     depth, boundary = _utils._get_depth_boundary(image.ndim, depth, "none")
+
+    if mode == "wrap":  # Fixes https://github.com/dask/dask-image/issues/242
+        boundary = "periodic"
+        mode = "constant"
 
     result = image.map_overlap(
         dispatch_convolve(image),
@@ -34,23 +33,23 @@ def convolve(image,
         weights=weights,
         mode=mode,
         cval=cval,
-        origin=origin
+        origin=origin,
     )
 
     return result
 
 
-@_utils._update_wrapper(scipy.ndimage.filters.correlate)
-def correlate(image,
-              weights,
-              mode='reflect',
-              cval=0.0,
-              origin=0):
+@_utils._update_wrapper(scipy.ndimage.correlate)
+def correlate(image, weights, mode="reflect", cval=0.0, origin=0):
     check_arraytypes_compatible(image, weights)
 
     origin = _utils._get_origin(weights.shape, origin)
     depth = _utils._get_depth(weights.shape, origin)
     depth, boundary = _utils._get_depth_boundary(image.ndim, depth, "none")
+
+    if mode == "wrap":  # Fixes https://github.com/dask/dask-image/issues/242
+        boundary = "periodic"
+        mode = "constant"
 
     result = image.map_overlap(
         dispatch_correlate(image),
@@ -61,7 +60,7 @@ def correlate(image,
         weights=weights,
         mode=mode,
         cval=cval,
-        origin=origin
+        origin=origin,
     )
 
     return result
