@@ -1,8 +1,8 @@
 import dask.array as da
 import numpy as np
 
-from . import _gaussian, _generic, _order
 from ..dispatch._dispatch_ndfilters import dispatch_threshold_local_mean
+from . import _gaussian, _generic, _order
 
 __all__ = [
     "threshold_local",
@@ -80,7 +80,7 @@ def threshold_local(image, block_size, method='gaussian', offset=0,
                                                mode=mode, cval=cval)
     elif method == 'gaussian':
         if param is None:
-            sigma = (da.asarray(block_size) - 1) / 6.0
+            sigma = (da.asarray(block_size, like=image._meta) - 1) / 6.0
         else:
             sigma = param
         thresh_image = _gaussian.gaussian_filter(image, sigma, mode=mode,
