@@ -1,42 +1,47 @@
 # -*- coding: utf-8 -*-
 
-
+import functools
+import operator
 import numbers
 
-import numpy
+import dask.array as da
+import numpy as np
 
-import dask.array
 
-
-def _get_freq_grid(shape, chunks, dtype=float):
+def _get_freq_grid(shape, chunks, axis, n, dtype=float):
     assert len(shape) == len(chunks)
 
     shape = tuple(shape)
-    dtype = numpy.dtype(dtype).type
+    dtype = np.dtype(dtype).type
 
     assert (issubclass(dtype, numbers.Real) and
             not issubclass(dtype, numbers.Integral))
 
-    freq_grid = [
-        dask.array.fft.fftfreq(s, chunks=c).astype(dtype)
-        for s, c in zip(shape, chunks)
-    ]
-    freq_grid = dask.array.meshgrid(*freq_grid, indexing="ij")
-    freq_grid = dask.array.stack(freq_grid)
+    axis = axis % len(shape)
+
+    freq_grid = []
+    for ax, (s, c) in enumerate(zip(shape, chunks)):
+        if axis == ax and n > 0:
+            f = da.fft.rfftfreq(n, chunks=c).astype(dtype)
+        else:
+            f = da.fft.fftfreq(s, chunks=c).astype(dtype)
+        freq_grid.append(f)
+
+    freq_grid = da.meshgrid(*freq_grid, indexing="ij", sparse=True)
 
     return freq_grid
 
 
-def _get_ang_freq_grid(shape, chunks, dtype=float):
-    dtype = numpy.dtype(dtype).type
+def _get_ang_freq_grid(shape, chunks, axis, n, dtype=float):
+    dtype = np.dtype(dtype).type
 
     assert (issubclass(dtype, numbers.Real) and
             not issubclass(dtype, numbers.Integral))
 
-    pi = dtype(numpy.pi)
+    pi = dtype(np.pi)
 
-    freq_grid = _get_freq_grid(shape, chunks, dtype=dtype)
-    ang_freq_grid = (2 * pi) * freq_grid
+    freq_grid = _get_freq_grid(shape, chunks, axis, n, dtype=dtype)
+    ang_freq_grid = tuple((2 * pi) * f for f in freq_grid)
 
     return ang_freq_grid
 
@@ -46,9 +51,9 @@ def _norm_args(a, s, n=-1, axis=-1):
         a = a.astype(float)
 
     if isinstance(s, numbers.Number):
-        s = numpy.array(a.ndim * [s])
-    elif not isinstance(s, dask.array.Array):
-        s = numpy.array(s)
+        s = np.array(a.ndim * [s])
+    elif not isinstance(s, da.Array):
+        s = np.array(s)
 
     if issubclass(s.dtype.type, numbers.Integral):
         s = s.astype(a.real.dtype)
@@ -59,9 +64,16 @@ def _norm_args(a, s, n=-1, axis=-1):
             "Shape of `s` must be 1-D and equal to the input's rank."
         )
 
-    if n != -1:
+    if n != -1 and a.shape[axis] != (n // 2 + 1):
         raise NotImplementedError(
-            "Currently `n` other than -1 is unsupported."
+            "In the case of real-valued images, it is required that "
+            "(n // 2 + 1) == image.shape[axis]."
         )
 
     return (a, s, n, axis)
+
+
+def _reshape_nd(arr, ndim, axis):
+    """Promote a 1d array to ndim with non-singleton size along axis."""
+    nd_shape = (1,) * axis + (arr.size,) + (1,) * (ndim - axis - 1)
+    return arr.reshape(nd_shape)
