@@ -1,19 +1,19 @@
 # -*- coding: utf-8 -*-
 
-__author__ = """John Kirkham"""
-__email__ = "kirkhamj@janelia.hhmi.org"
-
-
 import collections
 import functools
 import operator
+import warnings
+from dask import compute, delayed
 
-import numpy
-
-import dask.array
+import dask.array as da
+import dask.bag as db
+import dask.dataframe as dd
+import numpy as np
 
 from . import _utils
 from ._utils import _label
+from ._utils._find_objects import _array_chunk_location, _find_bounding_boxes, _find_objects
 
 __all__ = [
     "area",
@@ -78,14 +78,14 @@ def area(image, label_image=None, index=None):
     """
 
     if label_image is None:
-        return dask.array.prod(numpy.array([i for i in image.shape]))
+        return da.prod(np.array([i for i in image.shape]))
 
     else:
         image, label_image, index = _utils._norm_input_labels_index(
             image, label_image, index
         )
 
-        ones = dask.array.ones(
+        ones = da.ones(
             label_image.shape, dtype=bool, chunks=label_image.chunks
         )
 
@@ -128,8 +128,8 @@ def center_of_mass(image, label_image=None, index=None):
     # This only matters if index is some array.
     index = index.T
 
-    out_dtype = numpy.dtype([("com", float, (image.ndim,))])
-    default_1d = numpy.full((1,), numpy.nan, dtype=out_dtype)
+    out_dtype = np.dtype([("com", float, (image.ndim,))])
+    default_1d = np.full((1,), np.nan, dtype=out_dtype)
 
     func = functools.partial(
         _utils._center_of_mass, shape=image.shape, dtype=out_dtype
@@ -169,13 +169,13 @@ def extrema(image, label_image=None, index=None):
         image, label_image, index
     )
 
-    out_dtype = numpy.dtype([
+    out_dtype = np.dtype([
         ("min_val", image.dtype),
         ("max_val", image.dtype),
-        ("min_pos", numpy.dtype(numpy.int), image.ndim),
-        ("max_pos", numpy.dtype(numpy.int), image.ndim)
+        ("min_pos", np.dtype(int), image.ndim),
+        ("max_pos", np.dtype(int), image.ndim)
     ])
-    default_1d = numpy.zeros((1,), dtype=out_dtype)
+    default_1d = np.zeros((1,), dtype=out_dtype)
 
     func = functools.partial(
         _utils._extrema, shape=image.shape, dtype=out_dtype
@@ -193,15 +193,58 @@ def extrema(image, label_image=None, index=None):
         pos_nd = extrema_lbl[pos_key]
 
         if index.ndim == 0:
-            pos_nd = dask.array.squeeze(pos_nd)
+            pos_nd = da.squeeze(pos_nd)
         elif index.ndim > 1:
             pos_nd = pos_nd.reshape(
-                (int(numpy.prod(pos_nd.shape[:-1])), pos_nd.shape[-1])
+                (int(np.prod(pos_nd.shape[:-1])), pos_nd.shape[-1])
             )
 
         extrema_lbl[pos_key] = pos_nd
 
     result = tuple(extrema_lbl.values())
+
+    return result
+
+
+def find_objects(label_image):
+    """Return bounding box slices for each object labelled by integers.
+
+    Parameters
+    ----------
+    label_image : ndarray
+        Image features noted by integers.
+
+    Returns
+    -------
+    Dask dataframe
+        Each row respresents an indivdual integrer label. Columns contain the
+        slice information for the object boundaries in each dimension
+        (dimensions are named: 0, 1, ..., nd).
+
+    Notes
+    -----
+    You must have the optional dependency ``dask[dataframe]`` installed
+    to use the ``find_objects`` function.
+    """
+    if label_image.dtype.char not in np.typecodes['AllInteger']:
+        raise ValueError("find_objects only accepts integer dtype arrays")
+
+    block_iter = zip(
+        np.ndindex(*label_image.numblocks),
+        map(functools.partial(operator.getitem, label_image),
+            da.core.slices_from_chunks(label_image.chunks))
+    )
+
+    arrays = []
+    for block_id, block in block_iter:
+        array_location = _array_chunk_location(block_id, label_image.chunks)
+        arrays.append(delayed(_find_bounding_boxes)(block, array_location))
+
+    bag = db.from_sequence(arrays)
+    result = bag.fold(functools.partial(_find_objects, label_image.ndim), split_every=2).to_delayed()
+    meta = dd.utils.make_meta([(i, object) for i in range(label_image.ndim)])
+    result = delayed(compute)(result)[0]  # avoid the user having to call compute twice on result
+    result = dd.from_delayed(result, meta=meta, prefix="find-objects-", verify_meta=False)
 
     return result
 
@@ -289,17 +332,17 @@ def label(image, structure=None):
         How many objects were found.
     """
 
-    image = dask.array.asarray(image)
+    image = da.asarray(image)
 
-    labeled_blocks = numpy.empty(image.numblocks, dtype=object)
+    labeled_blocks = np.empty(image.numblocks, dtype=object)
 
     # First, label each block independently, incrementing the labels in that
     # block by the total number of labels from previous blocks. This way, each
     # block's labels are globally unique.
     block_iter = zip(
-        numpy.ndindex(*image.numblocks),
+        np.ndindex(*image.numblocks),
         map(functools.partial(operator.getitem, image),
-            dask.array.core.slices_from_chunks(image.chunks))
+            da.core.slices_from_chunks(image.chunks))
     )
     index, input_block = next(block_iter)
     labeled_blocks[index], total = _label.block_ndi_label_delayed(input_block,
@@ -307,15 +350,15 @@ def label(image, structure=None):
     for index, input_block in block_iter:
         labeled_block, n = _label.block_ndi_label_delayed(input_block,
                                                           structure)
-        block_label_offset = dask.array.where(labeled_block > 0,
-                                              total,
-                                              _label.LABEL_DTYPE.type(0))
+        block_label_offset = da.where(labeled_block > 0,
+                                      total,
+                                      _label.LABEL_DTYPE.type(0))
         labeled_block += block_label_offset
         labeled_blocks[index] = labeled_block
         total += n
 
     # Put all the blocks together
-    block_labeled = dask.array.block(labeled_blocks.tolist())
+    block_labeled = da.block(labeled_blocks.tolist())
 
     # Now, build a label connectivity graph that groups labels across blocks.
     # We use this graph to find connected components and then relabel each
@@ -324,7 +367,7 @@ def label(image, structure=None):
                                                 total)
     new_labeling = _label.connected_components_delayed(label_groups)
     relabeled = _label.relabel_blocks(block_labeled, new_labeling)
-    n = dask.array.max(relabeled)
+    n = da.max(relabeled)
 
     return (relabeled, n)
 
@@ -380,8 +423,8 @@ def labeled_comprehension(image,
         image, label_image, index
     )
 
-    out_dtype = numpy.dtype(out_dtype)
-    default_1d = numpy.full((1,), default, dtype=out_dtype)
+    out_dtype = np.dtype(out_dtype)
+    default_1d = np.full((1,), default, dtype=out_dtype)
 
     pass_positions = bool(pass_positions)
 
@@ -392,8 +435,8 @@ def labeled_comprehension(image,
         )
         args = (image, positions)
 
-    result = numpy.empty(index.shape, dtype=object)
-    for i in numpy.ndindex(index.shape):
+    result = np.empty(index.shape, dtype=object)
+    for i in np.ndindex(index.shape):
         lbl_mtch_i = (label_image == index[i])
         args_lbl_mtch_i = tuple(e[lbl_mtch_i] for e in args)
         result[i] = _utils._labeled_comprehension_func(
@@ -402,8 +445,8 @@ def labeled_comprehension(image,
 
     for i in range(result.ndim - 1, -1, -1):
         result2 = result[..., 0]
-        for j in numpy.ndindex(index.shape[:i]):
-            result2[j] = dask.array.stack(result[j].tolist(), axis=0)
+        for j in np.ndindex(index.shape[:i]):
+            result2[j] = da.stack(result[j].tolist(), axis=0)
         result = result2
     result = result[()][..., 0]
 
@@ -438,7 +481,7 @@ def maximum(image, label_image=None, index=None):
     )
 
     return labeled_comprehension(
-        image, label_image, index, numpy.max, image.dtype, image.dtype.type(0)
+        image, label_image, index, np.max, image.dtype, image.dtype.type(0)
     )
 
 
@@ -475,8 +518,8 @@ def maximum_position(image, label_image=None, index=None):
     if index.shape:
         index = index.flatten()
 
-    out_dtype = numpy.dtype([("pos", int, (image.ndim,))])
-    default_1d = numpy.zeros((1,), dtype=out_dtype)
+    out_dtype = np.dtype([("pos", int, (image.ndim,))])
+    default_1d = np.zeros((1,), dtype=out_dtype)
 
     func = functools.partial(
         _utils._argmax, shape=image.shape, dtype=out_dtype
@@ -488,7 +531,7 @@ def maximum_position(image, label_image=None, index=None):
     max_pos_lbl = max_pos_lbl["pos"]
 
     if index.shape == tuple():
-        max_pos_lbl = dask.array.squeeze(max_pos_lbl)
+        max_pos_lbl = da.squeeze(max_pos_lbl)
 
     return max_pos_lbl
 
@@ -520,10 +563,10 @@ def mean(image, label_image=None, index=None):
         image, label_image, index
     )
 
-    nan = numpy.float64(numpy.nan)
+    nan = np.float64(np.nan)
 
     mean_lbl = labeled_comprehension(
-        image, label_image, index, numpy.mean, numpy.float64, nan
+        image, label_image, index, np.mean, np.float64, nan
     )
 
     return mean_lbl
@@ -556,10 +599,10 @@ def median(image, label_image=None, index=None):
         image, label_image, index
     )
 
-    nan = numpy.float64(numpy.nan)
+    nan = np.float64(np.nan)
 
     return labeled_comprehension(
-        image, label_image, index, numpy.median, numpy.float64, nan
+        image, label_image, index, np.median, np.float64, nan
     )
 
 
@@ -591,7 +634,7 @@ def minimum(image, label_image=None, index=None):
     )
 
     return labeled_comprehension(
-        image, label_image, index, numpy.min, image.dtype, image.dtype.type(0)
+        image, label_image, index, np.min, image.dtype, image.dtype.type(0)
     )
 
 
@@ -625,8 +668,8 @@ def minimum_position(image, label_image=None, index=None):
     if index.shape:
         index = index.flatten()
 
-    out_dtype = numpy.dtype([("pos", int, (image.ndim,))])
-    default_1d = numpy.zeros((1,), dtype=out_dtype)
+    out_dtype = np.dtype([("pos", int, (image.ndim,))])
+    default_1d = np.zeros((1,), dtype=out_dtype)
 
     func = functools.partial(
         _utils._argmin, shape=image.shape, dtype=out_dtype
@@ -638,7 +681,7 @@ def minimum_position(image, label_image=None, index=None):
     min_pos_lbl = min_pos_lbl["pos"]
 
     if index.shape == tuple():
-        min_pos_lbl = dask.array.squeeze(min_pos_lbl)
+        min_pos_lbl = da.squeeze(min_pos_lbl)
 
     return min_pos_lbl
 
@@ -670,10 +713,10 @@ def standard_deviation(image, label_image=None, index=None):
         image, label_image, index
     )
 
-    nan = numpy.float64(numpy.nan)
+    nan = np.float64(np.nan)
 
     std_lbl = labeled_comprehension(
-        image, label_image, index, numpy.std, numpy.float64, nan
+        image, label_image, index, np.std, np.float64, nan
     )
 
     return std_lbl
@@ -707,7 +750,7 @@ def sum_labels(image, label_image=None, index=None):
     )
 
     sum_lbl = labeled_comprehension(
-        image, label_image, index, numpy.sum, numpy.float64, numpy.float64(0)
+        image, label_image, index, np.sum, np.float64, np.float64(0)
     )
 
     return sum_lbl
@@ -715,7 +758,8 @@ def sum_labels(image, label_image=None, index=None):
 
 def sum(image, label_image=None, index=None):
     """DEPRECATED FUNCTION. Use `sum_labels` instead."""
-    warnings.warn("DEPRECATED FUNCTION. Use `sum_labels` instead.", DeprecationWarning)
+    warnings.warn("DEPRECATED FUNCTION. Use `sum_labels` instead.",
+                  DeprecationWarning)
     return sum_labels(image, label_image=label_image, index=index)
 
 
@@ -746,10 +790,10 @@ def variance(image, label_image=None, index=None):
         image, label_image, index
     )
 
-    nan = numpy.float64(numpy.nan)
+    nan = np.float64(np.nan)
 
     var_lbl = labeled_comprehension(
-        image, label_image, index, numpy.var, numpy.float64, nan
+        image, label_image, index, np.var, np.float64, nan
     )
 
     return var_lbl
