@@ -1,20 +1,9 @@
 # -*- coding: utf-8 -*-
-
-
-from __future__ import division
-
 import warnings
 
-import numpy
-
 import dask
-import dask.array
-
-
-try:
-    from dask.array import blockwise as da_blockwise
-except ImportError:
-    from dask.array import atop as da_blockwise
+import dask.array as da
+import numpy as np
 
 
 def _norm_input_labels_index(image, label_image=None, index=None):
@@ -22,19 +11,19 @@ def _norm_input_labels_index(image, label_image=None, index=None):
     Normalize arguments to a standard form.
     """
 
-    image = dask.array.asarray(image)
+    image = da.asarray(image)
 
     if label_image is None:
-        label_image = dask.array.ones(
+        label_image = da.ones(
             image.shape, dtype=int, chunks=image.chunks,
         )
-        index = dask.array.ones(tuple(), dtype=int, chunks=tuple())
+        index = da.from_array(np.array(1, dtype=int))
     elif index is None:
         label_image = (label_image > 0).astype(int)
-        index = dask.array.ones(tuple(), dtype=int, chunks=tuple())
+        index = da.from_array(np.array(1, dtype=int))
 
-    label_image = dask.array.asarray(label_image)
-    index = dask.array.asarray(index)
+    label_image = da.asarray(label_image)
+    index = da.asarray(index)
 
     if index.ndim > 1:
         warnings.warn(
@@ -64,17 +53,17 @@ def _ravel_shape_indices(dimensions, dtype=int, chunks=None):
     """
 
     indices = [
-        dask.array.arange(
+        da.arange(
             0,
-            numpy.prod(dimensions[i:], dtype=dtype),
-            numpy.prod(dimensions[i + 1:], dtype=dtype),
+            np.prod(dimensions[i:], dtype=dtype),
+            np.prod(dimensions[i + 1:], dtype=dtype),
             dtype=dtype,
             chunks=c
         )
         for i, c in enumerate(chunks)
     ]
 
-    indices = da_blockwise(
+    indices = da.blockwise(
         _ravel_shape_indices_kernel, tuple(range(len(indices))),
         *sum([(a, (i,)) for i, a in enumerate(indices)], tuple()),
         dtype=dtype
@@ -88,9 +77,9 @@ def _argmax(a, positions, shape, dtype):
     Find original array position corresponding to the maximum.
     """
 
-    result = numpy.empty((1,), dtype=dtype)
+    result = np.empty((1,), dtype=dtype)
 
-    pos_nd = numpy.unravel_index(positions[numpy.argmax(a)], shape)
+    pos_nd = np.unravel_index(positions[np.argmax(a)], shape)
     for i, pos_nd_i in enumerate(pos_nd):
         result["pos"][0, i] = pos_nd_i
 
@@ -102,9 +91,9 @@ def _argmin(a, positions, shape, dtype):
     Find original array position corresponding to the minimum.
     """
 
-    result = numpy.empty((1,), dtype=dtype)
+    result = np.empty((1,), dtype=dtype)
 
-    pos_nd = numpy.unravel_index(positions[numpy.argmin(a)], shape)
+    pos_nd = np.unravel_index(positions[np.argmin(a)], shape)
     for i, pos_nd_i in enumerate(pos_nd):
         result["pos"][0, i] = pos_nd_i
 
@@ -116,14 +105,14 @@ def _center_of_mass(a, positions, shape, dtype):
     Find the center of mass for each ROI.
     """
 
-    result = numpy.empty((1,), dtype=dtype)
+    result = np.empty((1,), dtype=dtype)
 
-    positions_nd = numpy.unravel_index(positions, shape)
-    a_sum = numpy.sum(a)
+    positions_nd = np.unravel_index(positions, shape)
+    a_sum = np.sum(a)
 
-    a_wt_i = numpy.empty(a.shape)
+    a_wt_i = np.empty(a.shape)
     for i, pos_nd_i in enumerate(positions_nd):
-        a_wt_sum_i = numpy.multiply(a, pos_nd_i, out=a_wt_i).sum()
+        a_wt_sum_i = np.multiply(a, pos_nd_i, out=a_wt_i).sum()
         result["com"][0, i] = a_wt_sum_i / a_sum
 
     return result[0]
@@ -134,16 +123,16 @@ def _extrema(a, positions, shape, dtype):
     Find minimum and maximum as well as positions for both.
     """
 
-    result = numpy.empty((1,), dtype=dtype)
+    result = np.empty((1,), dtype=dtype)
 
-    int_min_pos = numpy.argmin(a)
-    int_max_pos = numpy.argmax(a)
+    int_min_pos = np.argmin(a)
+    int_max_pos = np.argmax(a)
 
     result["min_val"] = a[int_min_pos]
     result["max_val"] = a[int_max_pos]
 
-    min_pos_nd = numpy.unravel_index(positions[int_min_pos], shape)
-    max_pos_nd = numpy.unravel_index(positions[int_max_pos], shape)
+    min_pos_nd = np.unravel_index(positions[int_min_pos], shape)
+    max_pos_nd = np.unravel_index(positions[int_max_pos], shape)
     for i in range(len(shape)):
         result["min_pos"][0, i] = min_pos_nd[i]
         result["max_pos"][0, i] = max_pos_nd[i]
@@ -161,7 +150,7 @@ def _histogram(image,
     Also reformats the arguments.
     """
 
-    return numpy.histogram(image, bins, (min, max))[0]
+    return np.histogram(image, bins, (min, max))[0]
 
 
 @dask.delayed
@@ -177,7 +166,7 @@ def _labeled_comprehension_delayed(func,
     computation should not occur.
     """
 
-    result = numpy.empty((1,), dtype=out_dtype)
+    result = np.empty((1,), dtype=out_dtype)
 
     if a.size:
         if positions is None:
@@ -201,7 +190,7 @@ def _labeled_comprehension_func(func,
     Ensures the result is a proper Dask Array and the computation delayed.
     """
 
-    return dask.array.from_delayed(
+    return da.from_delayed(
         _labeled_comprehension_delayed(func, out_dtype, default, a, positions),
         (1,),
         out_dtype
