@@ -2,13 +2,12 @@
 
 import operator
 
-import numpy
+import dask
+import dask.array as da
+import numpy as np
 import scipy.ndimage
 import scipy.sparse
 import scipy.sparse.csgraph
-
-import dask
-import dask.array
 
 
 def _get_ndimage_label_dtype():
@@ -19,7 +18,7 @@ LABEL_DTYPE = _get_ndimage_label_dtype()
 
 
 def _get_connected_components_dtype():
-    a = numpy.empty((0, 0), dtype=int)
+    a = np.empty((0, 0), dtype=int)
     return scipy.sparse.csgraph.connected_components(a)[1].dtype
 
 
@@ -44,20 +43,20 @@ def relabel_blocks(block_labeled, new_labeling):
         The relabeled input array.
     """
     new_labeling = new_labeling.astype(LABEL_DTYPE)
-    relabeled = dask.array.map_blocks(operator.getitem,
-                                      new_labeling,
-                                      block_labeled,
-                                      dtype=LABEL_DTYPE,
-                                      chunks=block_labeled.chunks)
+    relabeled = da.map_blocks(operator.getitem,
+                              new_labeling,
+                              block_labeled,
+                              dtype=LABEL_DTYPE,
+                              chunks=block_labeled.chunks)
     return relabeled
 
 
 def _unique_axis(a, axis=0):
     """Find unique subarrays in axis in N-D array."""
-    at = numpy.ascontiguousarray(a.swapaxes(0, axis))
-    dt = numpy.dtype([("values", at.dtype, at.shape[1:])])
+    at = np.ascontiguousarray(a.swapaxes(0, axis))
+    dt = np.dtype([("values", at.dtype, at.shape[1:])])
     atv = at.view(dt)
-    r = numpy.unique(atv)["values"].swapaxes(0, axis)
+    r = np.unique(atv)["values"].swapaxes(0, axis)
     return r
 
 
@@ -86,26 +85,26 @@ def _across_block_label_grouping(face, structure):
 
     Examples
     --------
-    >>> face = numpy.array([[1, 1, 0, 2, 2, 0, 8],
+    >>> face = np.array([[1, 1, 0, 2, 2, 0, 8],
     ...                     [0, 7, 7, 7, 7, 0, 9]])
-    >>> structure = numpy.ones((3, 3), dtype=bool)
+    >>> structure = np.ones((3, 3), dtype=bool)
     >>> _across_block_label_grouping(face, structure)
     array([[1, 2, 8],
-           [2, 7, 9]], dtype=numpy.int32)
+           [2, 7, 9]], dtype=np.int32)
 
     This shows that 1-2 are connected, 2-7 are connected, and 8-9 are
     connected. The resulting graph is (1-2-7), (8-9).
     """
     common_labels = scipy.ndimage.label(face, structure)[0]
-    matching = numpy.stack((common_labels.ravel(), face.ravel()), axis=1)
+    matching = np.stack((common_labels.ravel(), face.ravel()), axis=1)
     unique_matching = _unique_axis(matching)
-    valid = numpy.all(unique_matching, axis=1)
+    valid = np.all(unique_matching, axis=1)
     unique_valid_matching = unique_matching[valid]
     common_labels, labels = unique_valid_matching.T
-    in_group = numpy.flatnonzero(numpy.diff(common_labels) == 0)
-    i = numpy.take(labels, in_group)
-    j = numpy.take(labels, in_group + 1)
-    grouped = numpy.stack((i, j), axis=0)
+    in_group = np.flatnonzero(np.diff(common_labels) == 0)
+    i = np.take(labels, in_group)
+    j = np.take(labels, in_group + 1)
+    grouped = np.stack((i, j), axis=0)
     return grouped
 
 
@@ -113,15 +112,13 @@ def _across_block_label_grouping_delayed(face, structure):
     """Delayed version of :func:`_across_block_label_grouping`."""
     _across_block_label_grouping_ = dask.delayed(_across_block_label_grouping)
     grouped = _across_block_label_grouping_(face, structure)
-    return dask.array.from_delayed(grouped,
-                                   shape=(2, numpy.nan),
-                                   dtype=LABEL_DTYPE)
+    return da.from_delayed(grouped, shape=(2, np.nan), dtype=LABEL_DTYPE)
 
 
 @dask.delayed
 def _to_csr_matrix(i, j, n):
     """Using i and j as coo-format coordinates, return csr matrix."""
-    v = numpy.ones_like(i)
+    v = np.ones_like(i)
     mat = scipy.sparse.coo_matrix((v, (i, j)), shape=(n, n))
     return mat.tocsr()
 
@@ -155,12 +152,12 @@ def label_adjacency_graph(labels, structure, nlabels):
         label j in the global volume, 0 everywhere else.
     """
     faces = _chunk_faces(labels.chunks, labels.shape)
-    all_mappings = [dask.array.empty((2, 0), dtype=LABEL_DTYPE, chunks=1)]
+    all_mappings = [da.empty((2, 0), dtype=LABEL_DTYPE, chunks=1)]
     for face_slice in faces:
         face = labels[face_slice]
         mapped = _across_block_label_grouping_delayed(face, structure)
         all_mappings.append(mapped)
-    all_mappings = dask.array.concatenate(all_mappings, axis=1)
+    all_mappings = da.concatenate(all_mappings, axis=1)
     i, j = all_mappings
     mat = _to_csr_matrix(i, j, nlabels + 1)
     return mat
@@ -184,7 +181,8 @@ def _chunk_faces(chunks, shape):
 
     Examples
     --------
-    >>> a = dask.array.arange(110, chunks=110).reshape((10, 11)).rechunk(5)
+    >>> import dask.array as da
+    >>> a = da.arange(110, chunks=110).reshape((10, 11)).rechunk(5)
     >>> chunk_faces(a.chunks, a.shape)
     [(slice(4, 6, None), slice(0, 5, None)),
      (slice(4, 6, None), slice(5, 10, None)),
@@ -194,7 +192,7 @@ def _chunk_faces(chunks, shape):
      (slice(5, 10, None), slice(4, 6, None)),
      (slice(5, 10, None), slice(9, 11, None))]
     """
-    slices = dask.array.core.slices_from_chunks(chunks)
+    slices = da.core.slices_from_chunks(chunks)
     ndim = len(shape)
     faces = []
     for ax in range(ndim):
@@ -228,9 +226,9 @@ def block_ndi_label_delayed(block, structure):
     label = dask.delayed(scipy.ndimage.label, nout=2)
     labeled_block, n = label(block, structure=structure)
     n = dask.delayed(LABEL_DTYPE.type)(n)
-    labeled = dask.array.from_delayed(labeled_block, shape=block.shape,
-                                      dtype=LABEL_DTYPE)
-    n = dask.array.from_delayed(n, shape=(), dtype=LABEL_DTYPE)
+    labeled = da.from_delayed(labeled_block, shape=block.shape,
+                              dtype=LABEL_DTYPE)
+    n = da.from_delayed(n, shape=(), dtype=LABEL_DTYPE)
     return labeled, n
 
 
@@ -242,5 +240,5 @@ def connected_components_delayed(csr_matrix):
     the number of components.
     """
     conn_comp = dask.delayed(scipy.sparse.csgraph.connected_components, nout=2)
-    return dask.array.from_delayed(conn_comp(csr_matrix, directed=False)[1],
-                                   shape=(numpy.nan,), dtype=CONN_COMP_DTYPE)
+    return da.from_delayed(conn_comp(csr_matrix, directed=False)[1],
+                           shape=(np.nan,), dtype=CONN_COMP_DTYPE)
