@@ -1,21 +1,12 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-
-from __future__ import absolute_import
-
 import pytest
-
 import numpy as np
-import scipy.ndimage.filters as sp_ndf
+import scipy.ndimage
 
-import dask
 import dask.array as da
-import dask.array.utils as dau
 
-import dask_image.ndfilters as da_ndf
-
-
-assert dask
+import dask_image.ndfilters
 
 
 @pytest.mark.parametrize(
@@ -29,8 +20,8 @@ assert dask
 @pytest.mark.parametrize(
     "da_func",
     [
-        da_ndf.prewitt,
-        da_ndf.sobel,
+        dask_image.ndfilters.prewitt,
+        dask_image.ndfilters.sobel,
     ]
 )
 def test_edge_func_params(da_func, err_type, axis):
@@ -44,8 +35,8 @@ def test_edge_func_params(da_func, err_type, axis):
 @pytest.mark.parametrize(
     "da_func",
     [
-        da_ndf.prewitt,
-        da_ndf.sobel,
+        dask_image.ndfilters.prewitt,
+        dask_image.ndfilters.sobel,
     ]
 )
 def test_edge_comprehensions(da_func):
@@ -57,8 +48,8 @@ def test_edge_comprehensions(da_func):
     l2s = [da_func(d[i]) for i in range(len(d))]
     l2c = [da_func(d[i])[None] for i in range(len(d))]
 
-    dau.assert_eq(np.stack(l2s), da.stack(l2s))
-    dau.assert_eq(np.concatenate(l2c), da.concatenate(l2c))
+    da.utils.assert_eq(np.stack(l2s), da.stack(l2s))
+    da.utils.assert_eq(np.concatenate(l2c), da.concatenate(l2c))
 
 
 @pytest.mark.parametrize(
@@ -75,8 +66,8 @@ def test_edge_comprehensions(da_func):
 @pytest.mark.parametrize(
     "da_func, sp_func",
     [
-        (da_ndf.prewitt, sp_ndf.prewitt),
-        (da_ndf.sobel, sp_ndf.sobel),
+        (dask_image.ndfilters.prewitt, scipy.ndimage.prewitt),
+        (dask_image.ndfilters.sobel, scipy.ndimage.sobel),
     ]
 )
 def test_edge_func_compare(da_func, sp_func, axis):
@@ -84,7 +75,7 @@ def test_edge_func_compare(da_func, sp_func, axis):
     a = np.arange(float(np.prod(s))).reshape(s)
     d = da.from_array(a, chunks=(5, 5, 6))
 
-    dau.assert_eq(
+    da.utils.assert_eq(
         sp_func(a, axis),
         da_func(d, axis)
     )
