@@ -1,21 +1,12 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-
-from __future__ import absolute_import
-
 import pytest
-
 import numpy as np
-import scipy.ndimage.filters as sp_ndf
+import scipy.ndimage
 
-import dask
 import dask.array as da
-import dask.array.utils as dau
 
-import dask_image.ndfilters as da_ndf
-
-
-assert dask
+import dask_image.ndfilters
 
 
 @pytest.mark.parametrize(
@@ -34,7 +25,7 @@ def test_uniform_filter_params(err_type, size, origin):
     d = da.from_array(a, chunks=(5, 7))
 
     with pytest.raises(err_type):
-        da_ndf.uniform_filter(d, size, origin=origin)
+        dask_image.ndfilters.uniform_filter(d, size, origin=origin)
 
 
 def test_uniform_shape_type():
@@ -46,13 +37,13 @@ def test_uniform_shape_type():
 
     assert all([(type(s) is int) for s in d.shape])
 
-    d2 = da_ndf.uniform_filter(d, size, origin=origin)
+    d2 = dask_image.ndfilters.uniform_filter(d, size, origin=origin)
 
     assert all([(type(s) is int) for s in d2.shape])
 
 
 def test_uniform_comprehensions():
-    da_func = lambda arr: da_ndf.uniform_filter(arr, 1, origin=0)  # noqa: E731
+    da_func = lambda arr: dask_image.ndfilters.uniform_filter(arr, 1, origin=0)  # noqa: E731, E501
 
     np.random.seed(0)
 
@@ -62,8 +53,8 @@ def test_uniform_comprehensions():
     l2s = [da_func(d[i]) for i in range(len(d))]
     l2c = [da_func(d[i])[None] for i in range(len(d))]
 
-    dau.assert_eq(np.stack(l2s), da.stack(l2s))
-    dau.assert_eq(np.concatenate(l2c), da.concatenate(l2c))
+    da.utils.assert_eq(np.stack(l2s), da.stack(l2s))
+    da.utils.assert_eq(np.concatenate(l2c), da.concatenate(l2c))
 
 
 @pytest.mark.parametrize(
@@ -76,13 +67,13 @@ def test_uniform_identity(size, origin):
     a = np.arange(140.0).reshape(10, 14)
     d = da.from_array(a, chunks=(5, 7))
 
-    dau.assert_eq(
-        d, da_ndf.uniform_filter(d, size, origin=origin)
+    da.utils.assert_eq(
+        d, dask_image.ndfilters.uniform_filter(d, size, origin=origin)
     )
 
-    dau.assert_eq(
-        sp_ndf.uniform_filter(a, size, origin=origin),
-        da_ndf.uniform_filter(d, size, origin=origin)
+    da.utils.assert_eq(
+        scipy.ndimage.uniform_filter(a, size, origin=origin),
+        dask_image.ndfilters.uniform_filter(d, size, origin=origin)
     )
 
 
@@ -102,7 +93,7 @@ def test_uniform_compare(size, origin):
     a = np.arange(float(np.prod(s))).reshape(s)
     d = da.from_array(a, chunks=(50, 55))
 
-    dau.assert_eq(
-        sp_ndf.uniform_filter(a, size, origin=origin),
-        da_ndf.uniform_filter(d, size, origin=origin)
+    da.utils.assert_eq(
+        scipy.ndimage.uniform_filter(a, size, origin=origin),
+        dask_image.ndfilters.uniform_filter(d, size, origin=origin)
     )
