@@ -1,12 +1,8 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-
-from __future__ import absolute_import
-
 import pytest
-
-import numpy
-import dask.array
+import numpy as np
+import dask.array as da
 
 from dask_image.ndmorph import _utils
 
@@ -16,12 +12,12 @@ from dask_image.ndmorph import _utils
     [
         (
             RuntimeError,
-            dask.array.ones([1, 2], dtype=bool, chunks=(1, 2,)),
-            dask.array.arange(2, dtype=bool, chunks=(2,))
+            da.ones([1, 2], dtype=bool, chunks=(1, 2,)),
+            da.arange(2, dtype=bool, chunks=(2,))
         ),
         (
             TypeError,
-            dask.array.arange(2, dtype=bool, chunks=(2,)),
+            da.arange(2, dtype=bool, chunks=(2,)),
             2.0
         ),
     ]
@@ -48,12 +44,12 @@ def test_errs__get_iterations(err_type, iterations):
     [
         (
             RuntimeError,
-            dask.array.arange(2, dtype=bool, chunks=(2,)),
-            dask.array.arange(1, dtype=bool, chunks=(2,))
+            da.arange(2, dtype=bool, chunks=(2,)),
+            da.arange(1, dtype=bool, chunks=(2,))
         ),
         (
             TypeError,
-            dask.array.arange(2, dtype=bool, chunks=(2,)),
+            da.arange(2, dtype=bool, chunks=(2,)),
             2.0
         ),
     ]
@@ -91,24 +87,24 @@ def test_errs__get_brute_force(err_type, brute_force):
     "expected, input, structure",
     [
         (
-            numpy.array([1, 1, 1], dtype=bool),
-            (dask.array.arange(10, chunks=(10,)) % 2).astype(bool),
+            np.array([1, 1, 1], dtype=bool),
+            (da.arange(10, chunks=(10,)) % 2).astype(bool),
             None
         ),
         (
-            numpy.array([[0, 1, 0], [1, 1, 1], [0, 1, 0]], dtype=bool),
-            (dask.array.arange(100, chunks=10).reshape(10, 10) % 2).astype(bool),  # noqa: E501
+            np.array([[0, 1, 0], [1, 1, 1], [0, 1, 0]], dtype=bool),
+            (da.arange(100, chunks=10).reshape(10, 10) % 2).astype(bool),  # noqa: E501
             None
         ),
         (
-            numpy.array([1, 1, 1], dtype=bool),
-            (dask.array.arange(10, chunks=(10,)) % 2).astype(bool),
-            numpy.array([1, 1, 1], dtype=int)
+            np.array([1, 1, 1], dtype=bool),
+            (da.arange(10, chunks=(10,)) % 2).astype(bool),
+            np.array([1, 1, 1], dtype=int)
         ),
         (
-            numpy.array([1, 1, 1], dtype=bool),
-            (dask.array.arange(10, chunks=(10,)) % 2).astype(bool),
-            numpy.array([1, 1, 1], dtype=bool)
+            np.array([1, 1, 1], dtype=bool),
+            (da.arange(10, chunks=(10,)) % 2).astype(bool),
+            np.array([1, 1, 1], dtype=bool)
         ),
     ]
 )
@@ -116,7 +112,7 @@ def test__get_structure(expected, input, structure):
     result = _utils._get_structure(input, structure)
 
     assert expected.dtype.type == result.dtype.type
-    assert numpy.array((expected == result).all())[()]
+    assert np.array((expected == result).all())[()]
 
 
 @pytest.mark.parametrize(
@@ -133,12 +129,12 @@ def test__get_iterations(expected, iterations):
 @pytest.mark.parametrize(
     "expected, a",
     [
-        (numpy.bool8, False),
-        (numpy.int_, 2),
-        (numpy.float64, 3.1),
-        (numpy.complex128, 1 + 2j),
-        (numpy.int16, numpy.int16(6)),
-        (numpy.uint32, numpy.arange(3, dtype=numpy.uint32)),
+        (np.bool8, False),
+        (np.int_, 2),
+        (np.float64, 3.1),
+        (np.complex128, 1 + 2j),
+        (np.int16, np.int16(6)),
+        (np.uint32, np.arange(3, dtype=np.uint32)),
     ]
 )
 def test__get_dtype(expected, a):
@@ -148,28 +144,28 @@ def test__get_dtype(expected, a):
 @pytest.mark.parametrize(
     "expected, input, mask",
     [
-        (True, dask.array.arange(2, dtype=bool, chunks=(2,)), None),
-        (True, dask.array.arange(2, dtype=bool, chunks=(2,)), True),
-        (False, dask.array.arange(2, dtype=bool, chunks=(2,)), False),
+        (True, da.arange(2, dtype=bool, chunks=(2,)), None),
+        (True, da.arange(2, dtype=bool, chunks=(2,)), True),
+        (False, da.arange(2, dtype=bool, chunks=(2,)), False),
         (
             True,
-            dask.array.arange(2, dtype=bool, chunks=(2,)),
-            numpy.bool8(True)
+            da.arange(2, dtype=bool, chunks=(2,)),
+            np.bool8(True)
         ),
         (
             False,
-            dask.array.arange(2, dtype=bool, chunks=(2,)),
-            numpy.bool8(False)
+            da.arange(2, dtype=bool, chunks=(2,)),
+            np.bool8(False)
         ),
         (
-            numpy.arange(2, dtype=bool),
-            dask.array.arange(2, dtype=bool, chunks=(2,)),
-            numpy.arange(2, dtype=bool)
+            np.arange(2, dtype=bool),
+            da.arange(2, dtype=bool, chunks=(2,)),
+            np.arange(2, dtype=bool)
         ),
         (
-            dask.array.arange(2, dtype=bool, chunks=(2,)),
-            dask.array.arange(2, dtype=bool, chunks=(2,)),
-            dask.array.arange(2, dtype=int, chunks=(2,))
+            da.arange(2, dtype=bool, chunks=(2,)),
+            da.arange(2, dtype=bool, chunks=(2,)),
+            da.arange(2, dtype=int, chunks=(2,))
         ),
     ]
 )
@@ -178,8 +174,8 @@ def test__get_mask(expected, input, mask):
 
     assert type(expected) == type(result)
 
-    if isinstance(expected, (numpy.ndarray, dask.array.Array)):
-        assert numpy.array((expected == result).all())[()]
+    if isinstance(expected, (np.ndarray, da.Array)):
+        assert np.array((expected == result).all())[()]
     else:
         assert expected == result
 
