@@ -1,16 +1,11 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-
-from __future__ import absolute_import
-
 import pytest
-
 import numpy as np
-import scipy.ndimage as spnd
+import scipy.ndimage
 
 import dask.array as da
-import dask.array.utils as dau
-import dask_image.ndmorph as da_ndm
+import dask_image.ndmorph
 
 
 @pytest.mark.parametrize(
@@ -50,7 +45,7 @@ def test_errs_binary_ops(funcname,
                          input,
                          structure,
                          origin):
-    da_func = getattr(da_ndm, funcname)
+    da_func = getattr(dask_image.ndmorph, funcname)
 
     with pytest.raises(err_type):
         da_func(
@@ -94,7 +89,7 @@ def test_errs_binary_ops_iter(funcname,
                               structure,
                               iterations,
                               origin):
-    da_func = getattr(da_ndm, funcname)
+    da_func = getattr(dask_image.ndmorph, funcname)
 
     with pytest.raises(err_type):
         da_func(
@@ -108,8 +103,10 @@ def test_errs_binary_ops_iter(funcname,
 @pytest.mark.parametrize(
     "funcname",
     [
+        "binary_closing",
         "binary_dilation",
         "binary_erosion",
+        "binary_opening",
     ]
 )
 @pytest.mark.parametrize(
@@ -157,7 +154,7 @@ def test_errs_binary_ops_expanded(funcname,
                                   border_value,
                                   origin,
                                   brute_force):
-    da_func = getattr(da_ndm, funcname)
+    da_func = getattr(dask_image.ndmorph, funcname)
 
     with pytest.raises(err_type):
         da_func(
@@ -269,8 +266,8 @@ def test_binary_ops(funcname,
                     input,
                     structure,
                     origin):
-    da_func = getattr(da_ndm, funcname)
-    sp_func = getattr(spnd, funcname)
+    da_func = getattr(dask_image.ndmorph, funcname)
+    sp_func = getattr(scipy.ndimage, funcname)
 
     da_result = da_func(
         input,
@@ -284,7 +281,7 @@ def test_binary_ops(funcname,
         origin=origin
     )
 
-    dau.assert_eq(sp_result, da_result)
+    da.utils.assert_eq(sp_result, da_result)
 
 
 @pytest.mark.parametrize(
@@ -348,8 +345,8 @@ def test_binary_ops_iter(funcname,
                          structure,
                          iterations,
                          origin):
-    da_func = getattr(da_ndm, funcname)
-    sp_func = getattr(spnd, funcname)
+    da_func = getattr(dask_image.ndmorph, funcname)
+    sp_func = getattr(scipy.ndimage, funcname)
 
     da_result = da_func(
         input,
@@ -365,14 +362,16 @@ def test_binary_ops_iter(funcname,
         origin=origin
     )
 
-    dau.assert_eq(sp_result, da_result)
+    da.utils.assert_eq(sp_result, da_result)
 
 
 @pytest.mark.parametrize(
     "funcname",
     [
+        "binary_closing",
         "binary_dilation",
         "binary_erosion",
+        "binary_opening",
     ]
 )
 @pytest.mark.parametrize(
@@ -490,8 +489,8 @@ def test_binary_ops_expanded(funcname,
                              border_value,
                              origin,
                              brute_force):
-    da_func = getattr(da_ndm, funcname)
-    sp_func = getattr(spnd, funcname)
+    da_func = getattr(dask_image.ndmorph, funcname)
+    sp_func = getattr(scipy.ndimage, funcname)
 
     da_result = da_func(
         input,
@@ -513,4 +512,4 @@ def test_binary_ops_expanded(funcname,
         brute_force=brute_force
     )
 
-    dau.assert_eq(sp_result, da_result)
+    da.utils.assert_eq(sp_result, da_result)
